Run installer as regular user
Running the installer as admin means that the default installation
folder is the admin's appdata folder, where the user that started the
installation cannot access it. The required admin privileges were
introduced to allow the qmk_driver_installer.exe to run.

Run the installer as a regular user and run qmk_driver_installer.exe
using shellexec to get a UAC prompt only when it is needed.